Move achievements and wallet to own folder
