+++ src/App.vue
@@ -37,8 +37,8 @@
 import {App} from "@/App.ts"
 import IgtSidebar from "@/components/util/sidebar/igt-sidebar-layout";
 import IgtTab from "@/components/util/igt-tab";
-import IgtAchievements from "@/components/igt-achievements";
-import IgtWallet from "@/components/igt-wallet";
+import IgtAchievements from "@/components/features/achievements/igt-achievements";
+import IgtWallet from "@/components/features/wallet/igt-wallet";
 import IgtNotifications from "@/components/util/igt-notifications";
 import IgtDeveloperPanel from "@/components/developer-panel/igt-developer-panel";
 import IgtSidebarCategory from "@/components/util/sidebar/igt-sidebar-category";
